[python] refactor datasource modules to decouple optional dependencies
