+++ paimon-python/pypaimon/read/datasource/ray_datasource.py
@@ -27,7 +27,7 @@ from typing import List, Optional, Iterable
 import pyarrow
 from packaging.version import parse
 import ray
-import torch
+from ray.data.datasource import Datasource
 
 from pypaimon.read.split import Split
 from pypaimon.read.table_read import TableRead
@@ -38,10 +38,6 @@ logger = logging.getLogger(__name__)
 # Ray version constants for compatibility
 RAY_VERSION_SCHEMA_IN_READ_TASK = "2.48.0"  # Schema moved from BlockMetadata to ReadTask
 RAY_VERSION_PER_TASK_ROW_LIMIT = "2.52.0"  # per_task_row_limit parameter introduced
-
-from ray.data.datasource import Datasource
-
-from torch.utils.data import Dataset, IterableDataset
 
 
 class RayDatasource(Datasource):
@@ -73,7 +69,6 @@ class RayDatasource(Datasource):
         if not self.splits:
             return 0
 
-        # Sum up file sizes from all splits
         total_size = sum(split.file_size for split in self.splits)
         return total_size if total_size > 0 else None
 
@@ -110,23 +105,18 @@ class RayDatasource(Datasource):
 
         per_task_row_limit = kwargs.get('per_task_row_limit', None)
 
-        # Validate parallelism parameter
         if parallelism < 1:
             raise ValueError(f"parallelism must be at least 1, got {parallelism}")
 
-        # Get schema for metadata
         if self._schema is None:
             self._schema = PyarrowFieldParser.from_paimon_schema(self.table_read.read_type)
 
-        # Adjust parallelism if it exceeds the number of splits
         if parallelism > len(self.splits):
             parallelism = len(self.splits)
             logger.warning(
                 f"Reducing the parallelism to {parallelism}, as that is the number of splits"
             )
 
-        # Store necessary information for creating readers in Ray workers
-        # Extract these to avoid serializing the entire self object in closures
         table = self.table_read.table
         predicate = self.table_read.predicate
         read_type = self.table_read.read_type
@@ -145,14 +135,11 @@ class RayDatasource(Datasource):
             from pypaimon.read.table_read import TableRead
             worker_table_read = TableRead(table, predicate, read_type)
 
-            # Read all splits in this chunk
             arrow_table = worker_table_read.to_arrow(splits)
 
-            # Return as a list to allow Ray to split into multiple blocks if needed
             if arrow_table is not None and arrow_table.num_rows > 0:
                 return [arrow_table]
             else:
-                # Return empty table with correct schema
                 empty_table = pyarrow.Table.from_arrays(
                     [pyarrow.array([], type=field.type) for field in schema],
                     schema=schema
@@ -229,118 +216,3 @@ class RayDatasource(Datasource):
             read_tasks.append(ReadTask(**read_task_kwargs))
 
         return read_tasks
-
-
-class TorchDataset(Dataset):
-    """
-    PyTorch Dataset implementation for reading Paimon table data.
-
-    This class enables Paimon table data to be used directly with PyTorch's
-    training pipeline, allowing for efficient data loading and batching.
-    """
-
-    def __init__(self, table_read: TableRead, splits: List[Split]):
-        """
-        Initialize TorchDataset.
-
-        Args:
-            table_read: TableRead instance for reading data
-            splits: List of splits to read
-        """
-        arrow_table = table_read.to_arrow(splits)
-        if arrow_table is None or arrow_table.num_rows == 0:
-            self._data = []
-        else:
-            self._data = arrow_table.to_pylist()
-
-    def __len__(self) -> int:
-        """
-        Return the total number of rows in the dataset.
-
-        Returns:
-            Total number of rows across all splits
-        """
-        return len(self._data)
-
-    def __getitem__(self, index: int):
-        """
-        Get a single item from the dataset.
-
-        Args:
-            index: Index of the item to retrieve
-
-        Returns:
-            Dictionary containing the row data
-        """
-        if not self._data:
-            return None
-
-        return self._data[index]
-
-
-class TorchIterDataset(IterableDataset):
-    """
-    PyTorch IterableDataset implementation for reading Paimon table data.
-
-    This class enables streaming data loading from Paimon tables, which is more
-    memory-efficient for large datasets. Data is read on-the-fly as needed,
-    rather than loading everything into memory upfront.
-    """
-
-    def __init__(self, table_read: TableRead, splits: List[Split]):
-        """
-        Initialize TorchIterDataset.
-
-        Args:
-            table_read: TableRead instance for reading data
-            splits: List of splits to read
-        """
-        self.table_read = table_read
-        self.splits = splits
-        # Get field names from read_type
-        self.field_names = [field.name for field in table_read.read_type]
-
-    def __iter__(self):
-        """
-        Iterate over the dataset, converting each OffsetRow to a dictionary.
-
-        Supports multi-worker data loading by partitioning splits across workers.
-        When num_workers > 0 in DataLoader, each worker will process a subset of splits.
-
-        Yields:
-            row data of dict type, where keys are column names
-        """
-        worker_info = torch.utils.data.get_worker_info()
-
-        if worker_info is None:
-            # Single-process data loading, iterate over all splits
-            splits_to_process = self.splits
-        else:
-            # Multi-process data loading, partition splits across workers
-            worker_id = worker_info.id
-            num_workers = worker_info.num_workers
-
-            # Calculate start and end indices for this worker
-            # Distribute splits evenly by slicing
-            total_splits = len(self.splits)
-            splits_per_worker = total_splits // num_workers
-            remainder = total_splits % num_workers
-
-            # Workers with id < remainder get one extra split
-            if worker_id < remainder:
-                start_idx = worker_id * (splits_per_worker + 1)
-                end_idx = start_idx + splits_per_worker + 1
-            else:
-                start_idx = worker_id * splits_per_worker + remainder
-                end_idx = start_idx + splits_per_worker
-
-            splits_to_process = self.splits[start_idx:end_idx]
-
-        worker_iterator = self.table_read.to_iterator(splits_to_process)
-
-        for offset_row in worker_iterator:
-            row_dict = {}
-            for i, field_name in enumerate(self.field_names):
-                value = offset_row.get_field(i)
-                row_dict[field_name] = value
-            yield row_dict
