+++ paimon-python/pypaimon/read/table_read.py
@@ -221,7 +221,7 @@ class TableRead:
         if override_num_blocks is not None and override_num_blocks < 1:
             raise ValueError(f"override_num_blocks must be at least 1, got {override_num_blocks}")
 
-        from pypaimon.read.datasource import RayDatasource
+        from pypaimon.read.datasource.ray_datasource import RayDatasource
         datasource = RayDatasource(self, splits)
         return ray.data.read_datasource(
             datasource,
@@ -234,11 +234,11 @@ class TableRead:
     def to_torch(self, splits: List[Split], streaming: bool = False) -> "torch.utils.data.Dataset":
         """Wrap Paimon table data to PyTorch Dataset."""
         if streaming:
-            from pypaimon.read.datasource import TorchIterDataset
+            from pypaimon.read.datasource.torch_dataset import TorchIterDataset
             dataset = TorchIterDataset(self, splits)
             return dataset
         else:
-            from pypaimon.read.datasource import TorchDataset
+            from pypaimon.read.datasource.torch_dataset import TorchDataset
             dataset = TorchDataset(self, splits)
             return dataset
